feat: implement CV component structure with sections for header, summary, skills, achievements, experience, education, personal projects, courses, and footer
