+++ apps/cv/src/components/cv/components/Courses.tsx
@@ -1,0 +1,16 @@
+export function Courses() {
+  return (
+    <section className="mb-4">
+      <h3 className="text-lg font-bold mb-3 pb-1 border-b-2 border-black">
+        COURSES
+      </h3>
+      <ul className="text-sm space-y-1">
+        <li>• Microsoft Certified: Azure AI Fundamentals</li>
+        <li>• Microsoft Certified: Azure Fundamentals</li>
+        <li>• AI Devs 2 & 3 - AI integration and building agents</li>
+        <li>• Modern Frontend - Next.js, React, GraphQL and Typescript</li>
+        <li>• Architecture on Frontend (Architektura na Froncie)</li>
+      </ul>
+    </section>
+  );
+}
