+++ apps/cv/src/components/cv/components/Education.tsx
@@ -1,0 +1,20 @@
+export function Education() {
+  return (
+    <section className="mb-4">
+      <h3 className="text-lg font-bold mb-3 pb-1 border-b-2 border-black">
+        EDUCATION
+      </h3>
+      <div className="mb-4">
+        <h4 className="font-bold text-base">
+          Bachelor of Science in Computer Science
+        </h4>
+        <div className="flex justify-between items-center">
+          <div className="font-semibold text-sm">
+            University of Warmia and Mazury in Olsztyn
+          </div>
+          <div className="text-sm text-gray-600">2017 - 2021</div>
+        </div>
+      </div>
+    </section>
+  );
+}
